requirements.txt
streamlit
google-generativeai
beautifulsoup4
requests
pandas
janome
openpyxl
google-api-python-client
